Moved from Vue CLI to Vite
