+++ index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width,initial-scale=1.0">
+      <title>Lecteur de tickets</title>
+  </head>
+  <body>
+    <div id="app"></div>
+
+    <script src="src/main.ts" type="module"></script>
+  </body>
+</html>
