+++ src/router/index.ts
@@ -4,7 +4,7 @@ import Home from '@/views/Home.vue'
 const routes: Array<RouteRecordRaw> = []
 
 const router = createRouter({
-  history: createWebHistory(process.env.BASE_URL),
+  history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
