+++ vite.config.ts
@@ -1,0 +1,11 @@
+import {defineConfig} from "vite";
+import vue from "@vitejs/plugin-vue";
+
+export default defineConfig({
+  plugins: [vue()],
+  resolve: {
+    alias: [
+      {find: "@", replacement: "/src"},
+    ]
+  }
+})
